# apps/server/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditListResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "entries"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "password",
          "username"
        ],
        "type": "object"
      },
...
      "OKResponse": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "csrfToken": {
            "description": "Send as X-CSRF-Token on POST/PATCH",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "csrfToken",
          "user"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "in": "cookie",
        "name": "fbw_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Unsafe methods need the X-CSRF-Token header returned by /api/auth/login or /api/me.",
    "title": "Find Bugs Web API",
    "version": "dev"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/audit": {
      "get": {
        "operationId": "listAudit",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditListResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Audit log of VPN/Splunk actions",
        "tags": [
          "users"
        ],
        "x-roles": [
          "engineer"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Log in",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OKResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/environments": {
      "get": {
        "operationId": "environments",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvironmentsResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Form options",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs": {
      "get": {
        "operationId": "listJobs",
        "parameters": [
          {
            "in": "query",
            "name": "environment",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "transactionId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "beforeId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List jobs (QA: own only)",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "submitJob",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitJobResponse"
                }
              }
            },
            "description": "OK"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitJobResponse"
                }
              }
            },
            "description": "Accepted"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Submit an investigation",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Job detail and result",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{id}/cancel": {
      "post": {
        "operationId": "cancelJob",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Cancel a pending job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/me": {
      "get": {
        "operationId": "me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Current user",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}